FSM, no viewer
Moved wait_next

Fix

+++ src/ccnc.h
@@ -9,9 +9,10 @@
 #include <math.h>
 #include <assert.h>
 #include <string.h>
+#include <machine.h>
 
 typedef double data_t;
-typedef size_t index_t;
+typedef uint32_t index_t;
